requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
dj-database-url==2.0.0
Django==4.2.1
django-cloudinary-storage==0.3.0
django-jazzmin==2.6.0
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
psycopg
pycodestyle==2.10.0
python-decouple==3.8
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
django-ninja
